add router delete and put
